feat: separate to components
