+++ src/routes/graphql/types/types.ts
@@ -1,0 +1,9 @@
+import { GraphQLString } from 'graphql';
+
+export interface ResolveArgs {
+  id: string;
+}
+
+export const typeArgs = {
+  id: { type: GraphQLString },
+};
